Move everything except scripts into utils
