+++ src/magic_tiler/magic_tiler.py
@@ -4,8 +4,8 @@ import pprint
 import click
 
 import magic_tiler
-from magic_tiler import dtos
 from magic_tiler import sway
+from magic_tiler.utils import dtos
 from magic_tiler.utils import subprocess_runner
 
 # maps from verbosity level to log levels
