+++ src/magic_tiler/utils/dtos.py
