+++ src/magic_tiler/utils/layout.py
@@ -3,7 +3,7 @@ import logging
 import pprint
 from typing import Dict, List, NamedTuple
 
-from magic_tiler import dtos
+from magic_tiler.utils import dtos
 from magic_tiler.utils import interfaces
