+++ src/magic_tiler/utils/sway.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Tuple
 
 import i3ipc
 
-from magic_tiler import dtos
+from magic_tiler.utils import dtos
 from magic_tiler.utils import interfaces
 
 """We need to sleep for a short time since processes take time to start.
