+++ tests/test_layout.py
@@ -2,7 +2,7 @@ from typing import Dict, List, NamedTuple
 
 import pytest
 
-from magic_tiler import dtos
+from magic_tiler.utils import dtos
 from magic_tiler.utils import interfaces
 from magic_tiler.utils import layout
